+++ changelog.md
@@ -1,1 +1,2 @@
-fixed cauldron dye 
+- added a sign blacklist for pixel consistent signs
+- fixed 2 ghasts fireballs mixins not being applied causing amendments particles to be missing